backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
python-dotenv==1.0.1
sentence-transformers==3.2.1
faiss-cpu==1.8.0; platform_system != "Windows"
numpy==1.26.4; python_version < "3.13"
numpy>=2.1.0; python_version >= "3.13"
groq==0.11.0
httpx==0.27.2
